+++ src/modules/articleParser/types/index.ts
@@ -1,0 +1,5 @@
+export interface Article {
+  title: string;
+  image?: string;
+  content: string;
+}
