+++ src/modules/auth/types/index.ts
@@ -1,0 +1,5 @@
+export interface User {
+  id: string;
+  email: string;
+  name: string;
+}
