+++ src/modules/feedParser/services/feedParser.service.ts
@@ -1,6 +1,7 @@
 import type { FastifyInstance } from "fastify";
 import Parser, { type Item } from "rss-parser";
 import { getFeedFromDB, saveFeedToDB } from "./mongo.service";
+import { Feed } from "../types";
 
 interface CustomItem extends Item {
   "content:encoded"?: string;
@@ -8,22 +9,22 @@ interface CustomItem extends Item {
 
 const parser = new Parser();
 
-export interface NewsItem {
-  title: string;
-  link: string;
-  image: string;
-  pubDate: string;
-  contentSnippet: string;
-  content: string;
-  isoDate: string;
-  [key: string]: unknown;
-}
+// export interface NewsItem {
+//   title: string;
+//   link: string;
+//   image: string;
+//   pubDate: string;
+//   contentSnippet: string;
+//   content: string;
+//   isoDate: string;
+//   [key: string]: unknown;
+// }
 
-export interface Feed {
-  title: string;
-  items: NewsItem[];
-  [key: string]: unknown;
-}
+// export interface Feed {
+//   title: string;
+//   items: NewsItem[];
+//   [key: string]: unknown;
+// }
 
 export async function parseFeed(
   fastify: FastifyInstance,
